Split large config test into multiple focused files
